refactor: onboarding form
