+++ packages/api/src/routers/auth.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 import { protectedProcedure, publicProcedure, router } from "../trpc";
 
 export const authRouter = router({
-  getCurrentSession: protectedProcedure.query(({ ctx }) => ctx.auth),
+  getCurrentSession: publicProcedure.query(({ ctx }) => ctx.auth),
   getCurrentUser: protectedProcedure.query(async ({ ctx }) => {
     const { data: user } = await ctx.supabase
       .from("users")
